Mark scale choices with minimum scale change difference
Add a method to find best set of scale choices among the available
alternatives in order to minimize the number of note changes between
scales. Mark best choices in the output with "star" characters.
This change also changes the way we output the scales by always using
C as the first displayed note, instead of the scale root.